pyproject.toml
[project]
name = "mcp-server-snowflake"
version = "0.4.0"
description = "MCP server for Snowflake"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  {name = "Carlos Serrano"},
  {name = "Jason Summer"},
  {name = "Tyler White"},
]
dependencies = [
    "fastmcp>=2.8.1",
    "mcp[cli]>=1.9.4",
    "pydantic>=2.11.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "snowflake-connector-python>=3.15.0",
    "snowflake-core>=1.0.0",
]

[dependency-groups]
dev = [
    "fast-agent-mcp>=0.2.24",
    "pre-commit>=4.2.0",
    "pyright>=1.1.402",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.8",
]

[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[project.scripts]
mcp-server-snowflake = "mcp_server_snowflake:main"

[tool.codespell]
ignore-words-list = "inout"
